chore(cluster): align cluster_options with ADR-003/005
